requirements.txt
fastapi
uvicorn
pyjwt
pytest
pytest-cov
ruff
httpx
requests
